% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/autofun-package.R
\docType{package}
\name{autofun-package}
\alias{autofun-package}
\title{autofun: Automatic Generation of R Functions}
\description{
Automatically generate an R funtion using source code contained
in an R script.
}
\details{
\tabular{ll}{
Package: \tab autofun   \cr
Type:    \tab Package   \cr
Version: \tab 0.0.0.9000\cr
Date:    \tab 2015-07-13\cr
License: \tab GPL (>= 2)\cr
}
}
\author{
Author and Maintainer: Alessandro Samuel-Rosa
\email{alessandrosamuelrosa@gmail.com}
}

