% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/autofun.R
\name{autofun}
\alias{autofun}
\title{Automatic Generation of R Functions}
\usage{
autofun(fun.name, read.file, write.file)
}
\arguments{
\item{fun.name}{Name of the function.}

\item{read.file}{A connection, or a character string naming the file that
contains the R source code that will constitute the body of the
function.}

\item{write.file}{A connection, or a character string naming the file to
write the function to.}
}
\description{
Automatically generate an R funtion using source code contained in an R
script.
}
\details{
The idea behind \code{autofun} is to use a standalone R script to create an
R function. This is done parsing the R script and using the returned
expression as the function body.

A function \code{foo} returned by \code{autofun} gets all the arguments from
the working environment. This is done by evaluating the function with
\code{eval(foo())}. Such a function can return several objects.
}
\author{
Alessandro Samuel-Rosa \email{alessandrosamuelrosa@gmail.com}
}
\seealso{
\code{\link{base}{parse}}, \code{\link{utils}{capture.output}},
\code{\link{base}{eval}}, \code{\link{base}{expression}}
}

